Fix up docs after I tried to install it
I ran into a couple issues when trying to get this running
for the first time so I have revised the README to have
a very explicit step by step set of instructions to get
the code to the point of compiling.

I also fixed up the LICENSE file a little bit so that it
is LICENSE.txt and has "Modus Create Inc." in it.